+++ tomcat/src/main/java/com/techcourse/Application.java
@@ -1,8 +1,5 @@
 package com.techcourse;
 
-import com.techcourse.controller.LoginController;
-import com.techcourse.controller.RegisterController;
-import org.apache.catalina.startup.Controllers;
 import org.apache.catalina.startup.Server;
 import org.apache.catalina.startup.Tomcat;
 import org.apache.catalina.startup.WAS;
@@ -10,11 +7,7 @@ import org.apache.catalina.startup.WAS;
 public class Application {
     public static void main(String[] args) {
         Server server = new Tomcat();
-        Controllers controllers = new Controllers(
-                new LoginController("/login"),
-                new RegisterController("/register")
-        );
 
-        new WAS(server, controllers).start();
+        new WAS(server).start();
     }
 }
