+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -7,7 +7,7 @@ import com.techcourse.model.User;
 import com.techcourse.session.Session;
 import com.techcourse.session.SessionManager;
 import java.util.Map;
-import org.apache.coyote.http11.HttpStatus;
+import org.apache.coyote.HttpStatus;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
