+++ tomcat/src/main/java/org/apache/coyote/http11/request/Http11Reader.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import java.io.BufferedReader;
 import java.io.IOException;
